hirewise-backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# AI Integration
google-generativeai==0.3.1

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2
spacy==3.7.2

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
